+++ app/feedback/model.go
@@ -8,7 +8,7 @@ import (
 
 //Idea represents an idea on a tenant board
 type Idea struct {
-	ID          int64     `json:"id"`
+	ID          int       `json:"id"`
 	Title       string    `json:"title"`
 	Description string    `json:"description"`
 	CreatedOn   time.Time `json:"createdOn"`
@@ -17,7 +17,7 @@ type Idea struct {
 
 //Comment represents an user comment on an idea
 type Comment struct {
-	ID        int64     `json:"id"`
+	ID        int       `json:"id"`
 	Content   string    `json:"content"`
 	CreatedOn time.Time `json:"createdOn"`
 	User      app.User  `json:"user"`
